[python] Add `verbose` deprecation notice and `util.py`
- Add a deprecation notice for the optional `verbose` argument
- Create `tools/util.py` to contain technical helpers such as
  the deprecation warning used in various submodules.
