+++ pypmc/mix_adapt/hierarchical_test.py
@@ -29,7 +29,7 @@ class TestHierarchical(unittest.TestCase):
 
     def test_prune(self):
         h = Hierarchical(self.input_components1, self.initial_guess1)
-        h.run(verbose=True)
+        h.run()
         sol = h.g
 
         # only one component should survive and have weight 1.0
@@ -46,7 +46,7 @@ class TestHierarchical(unittest.TestCase):
 
     def test_cluster(self):
         h = Hierarchical(self.input_components2, self.initial_guess2)
-        h.run(verbose=True)
+        h.run()
         sol = h.g
 
         # both components should survive and have equal weight
