+++ pypmc/mix_adapt/pmc.pyx
@@ -429,11 +429,11 @@ class PMC(object):
             .. math::
                 \| L_t - L_{t-1} \| < \epsilon_a .
 
-        :param verbose:
-
-            Output status information after each update.
-
         '''
+        if verbose:
+            from pypmc.tools.util import depr_warn_verbose
+            depr_warn_verbose( __name__)
+
         old_K = None
 
         for i in range(1, iterations + 1):
