+++ pypmc/tools/__init__.py
@@ -5,4 +5,4 @@
 from ._history import History
 from ._partition import partition, patch_data
 from ._plot import plot_mixture, plot_responsibility
-from . import indicator, convergence
+from . import indicator, convergence, util
